cocos2d-x3.0rc0の64bit対応による暗黙型変換の警告対応

+++ Classes/quest/SRPGMapLayer.cpp
@@ -375,7 +375,7 @@ void SRPGMapLayer::executeMapIndex(MapIndex mapIndex)
         {
             // 移動アニメーション作成。移動後にマップカーソルを初期化
 
-            int movePointSize = list.size();
+            long movePointSize = list.size();
             auto moveArray = Vector<FiniteTimeAction*>();
             for (auto mapIndex : list)
             {
